refactor: split auth and user actions into separate files
