+++ src/app/store/effects/auth.effects.ts
@@ -5,6 +5,7 @@ import { catchError, map, switchMap, tap } from 'rxjs/operators';
 import { of } from 'rxjs';
 
 import * as AuthActions from '../actions/auth.actions';
+import * as UserActions from '../actions/user.actions';
 import { ApiService } from 'src/app/services/api.service';
 
 @Injectable()
@@ -37,11 +38,11 @@ export class AuthEffects {
   /** Effect for loading assessments action */
   public loadAssessments$ = createEffect(() =>
     this.actions$.pipe(
-      ofType(AuthActions.loadAssessments),
+      ofType(UserActions.loadAssessments),
       switchMap(() =>
         this.apiService.getUserAssessments().pipe(
-          map(assessments => AuthActions.loadAssessmentsSuccess({ assessments })),
-          catchError(error => of(AuthActions.loadAssessmentsFailure({ error })))
+          map(assessments => UserActions.loadAssessmentsSuccess({ assessments })),
+          catchError(error => of(UserActions.loadAssessmentsFailure({ error })))
         )
       )
     )
@@ -50,11 +51,11 @@ export class AuthEffects {
   /** Effect for loading users action */
   public loadUsers$ = createEffect(() =>
     this.actions$.pipe(
-      ofType(AuthActions.loadUsers),
+      ofType(UserActions.loadUsers),
       switchMap(() =>
         this.apiService.getUsers().pipe(
-          map(users => AuthActions.loadUsersSuccess({ users })),
-          catchError(error => of(AuthActions.loadUsersFailure({ error })))
+          map(users => UserActions.loadUsersSuccess({ users })),
+          catchError(error => of(UserActions.loadUsersFailure({ error })))
         )
       )
     )
@@ -63,11 +64,11 @@ export class AuthEffects {
   /** Effect for loading a specific assessment graph */
   public loadAssessmentGraph$ = createEffect(() =>
     this.actions$.pipe(
-      ofType(AuthActions.loadAssessmentGraph),
+      ofType(UserActions.loadAssessmentGraph),
       switchMap(({ assessmentId }) =>
         this.apiService.getAssessmentGraph(assessmentId).pipe(
-          map(graphData => AuthActions.loadAssessmentGraphSuccess({ graphData })),
-          catchError(error => of(AuthActions.loadAssessmentGraphFailure({ error })))
+          map(graphData => UserActions.loadAssessmentGraphSuccess({ graphData })),
+          catchError(error => of(UserActions.loadAssessmentGraphFailure({ error })))
         )
       )
     )
